chore: lots of organization on the GUI systems
