+++ Content/Buffs/Rage.cs
@@ -1,4 +1,5 @@
 ﻿using PathOfTerraria.Core;
+using PathOfTerraria.Core.Events;
 
 namespace PathOfTerraria.Content.Buffs;
