+++ Content/GUI/Utilities/UIPanelTab.cs
@@ -1,7 +1,7 @@
 ﻿using Terraria.GameContent.UI.Elements;
 using Terraria.Localization;
 
-namespace PathOfTerraria.Content.GUI;
+namespace PathOfTerraria.Content.GUI.Utilities;
 
 public class UIPanelTab(string name, LocalizedText text, float textScale = 1, bool large = false)
 	: UITextPanel<LocalizedText>(text, textScale, large)
