+++ Content/GUI/Utilities/UiStateRefresher.cs
@@ -1,12 +1,6 @@
 using PathOfTerraria.Core.Loaders.UILoading;
-using PathOfTerraria.Core.Systems.ModPlayers;
-using System.Collections.Generic;
-using PathOfTerraria.Content.Items.Gear;
-using Terraria.Audio;
-using Terraria.ID;
-using Terraria.UI;
 
-namespace PathOfTerraria.Content.GUI;
+namespace PathOfTerraria.Content.GUI.Utilities;
 
 /// <summary>
 /// This is used for recalculating the bounding box for the UI on resizes.
