+++ Core/Systems/Questing/QuestingEventsHandlers/QuestingEventsPlayer.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PathOfTerraria.Core.Systems.Questing.QuestingEventHandlers;
+﻿namespace PathOfTerraria.Core.Systems.Questing.QuestingEventsHandlers;
 internal class QuestingEventsPlayer : ModPlayer
 {
